fix(player_strategy)!: choose_card_to_flip does not take a hand as argument
Before this commit, the choose_card_to_flip function took a hand as
argument. This was not correct, since it takes the cards in front
of the player as argument, so no need to pass the hand.

BRAKING CHANGE: the choose_card_to_flip function now only takes the
player (and the game_status) as argument and not the hand.